+++ plugin/src/net/groboclown/idea/p4ic/v2/ui/alerts/RetryAuthenticationFailedHandler.java
@@ -43,7 +43,7 @@ public class RetryAuthenticationFailedHandler extends AbstractErrorHandler {
 
     @Override
     public void handleError(@NotNull final Date when) {
-        LOG.warn("P4ServerName refused authentication too many times");
+        LOG.warn("P4ServerName refused authentication too many times for server " + config.getServerId());
 
         DistinctDialog.performOnDialog(
                 DistinctDialog.key(this, config.getServerName(), config.getUsername()),